pyproject.toml
[tool.poetry]
name = "flask-demo"
version = "0.1.0"
description = ""
authors = ["顾平安 <g332920287@foxmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
flask = "3.0.2"
toml = "^0.10.2"
requests = "^2.32.2"
flask-wtf = "^1.2.1"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.1"
flask-mysqldb = "^2.0.0"
aiomysql = "^0.2.0"

[[tool.poetry.source]]
name = "tencent-pypi"
url = "http://mirrors.cloud.tencent.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
